fix(contracts): cambiar contratos financieros de float a int (ADR-018)
CIERRE DE P2-005:
Varios contratos financieros (interfaces, métodos y docblocks) todavía
declaraban `float` como tipo de retorno para montos monetarios, contradiciendo
directamente ADR-018 (Integer CLP) y la migración completa a INTEGER.

Esto generaba inconsistencia arquitectónica: los datos eran INTEGER en la BD,
pero los contratos prometían float, lo que podía causar confusiones,
type-juggling silencioso y errores en consumidores de estos contratos.

CAMBIOS IMPLEMENTADOS:
✅ Account.php: `balance(): float` → `balance(): int`
✅ PaymentQueryServiceInterface.php:
   - `getWaiterTipsInSession(): float` → `: int`
   - `@return float` → `@return int` (docblocks)
   - `array{total_amount: float, total_tips: float}` → `int` (PHPDoc shapes)
   - `Collection<string, float>` → `Collection<string, int>`
✅ PaymentQueryService.php: Implementación alineada con `: int`
✅ CashSession.php: Docblock `@return float` → `@return int`

IMPACTO:
✅ Coherencia total entre esquema de BD (INTEGER), contratos (int) y
   implementaciones (int) en todo el dominio financiero.
✅ Prevención de type-juggling silencioso en consumidores de contratos.
✅ Cumplimiento estricto de ADR-018 (Money Contract) y ADR-011 (Integer CLP).

VERIFICACIÓN:
✅ Tests de Accounting y Payments pasando (71 tests, 179 assertions).
✅ Cero float restantes en contratos financieros.

Refs: P2-005 Contratos contables float, ADR-018 Money Contract, ADR-011 Integer CLP

+++ app/Modules/Accounting/Domain/Entities/Account.php
@@ -72,7 +72,7 @@ class Account extends Model
      * Para cuentas con saldo normal crédito (Liability, Equity, Revenue):
      *   Balance = SUM(credits) - SUM(debits)
      */
-    public function balance(): float
+    public function balance(): int
     {
         $debits = $this->ledgerEntries()->sum('debit_amount');
         $credits = $this->ledgerEntries()->sum('credit_amount');
